minimal-example/backend/requirements.txt
# Django框架和核心依赖
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0

# 数据库
psycopg2-binary==2.9.9

# 过滤和查询
django-filter==23.5
django-cors-headers==4.4.0

# 数据处理
pandas==2.1.4
numpy==1.24.4

# API文档
drf-yasg==1.21.7

# 时区支持
pytz==2024.1

# 环境变量管理
python-decouple==3.8

# 文件上传和处理
Pillow==10.1.0

# 缓存
redis==5.0.1
django-redis==5.4.0

# 任务队列（可选）
celery==5.3.4

# 开发工具
pytest==7.4.3
pytest-django==4.7.0
requests==2.31.0